requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-restx==1.3.0
werkzeug>=3.0.0
google-generativeai==0.3.2
pydantic==2.5.2
python-dotenv==1.0.0
flask-sqlalchemy==3.1.1
sqlalchemy==2.0.23
typing-extensions==4.8.0
requests==2.31.0
pymysql==1.1.0
cryptography==41.0.7
